pyproject.toml
# pyproject.toml
[project]
name = "chat-backend"
version = "0.1.0"
description = "Replicación del backend con FastAPI"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  # Core FastAPI and ASGI
  "fastapi>=0.103.1",
  "uvicorn>=0.23.2",
  "websockets>=11.0.3",
  "asgi-lifespan>=2.1.0",
  "gunicorn>=21.2.0",
  "uvloop==0.19.0", # specific version for compatibility
  "httptools==0.6.1", # specific version for compatibility

  # Database and ORM
  "alembic>=1.12.0",
  "sqlalchemy>=2.0.20",
  "asyncpg>=0.28.0",
  "redis<5.0.0", # specific upper bound
  "sqladmin>=0.16.0", # for admin interface
  "psycopg2-binary ~=2.9",

  # Authentication and Security
  "pydantic[email]>=2.3.0",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt>=4.1.0",
  "pyjwt>=2.8.0",
  "itsdangerous>=2.1.2", # dependency for sqladmin/other auth
  "authlib[starlette]>=1.2.0,<2.0.0",

  # HTTP Clients
  "httpx>=0.25.0", # for making HTTP requests
  "requests>=2.31.0", # also common for HTTP requests (ensure no redundancy with httpx if not both used)

  # Mercado Pago
  "mercadopago>=2.3.0",

  # Logging
  "loguru>=0.7.3",

  # Google Auth para OAuth2
  "google-auth>=2.23.0",

  # AI/NLP Related Libraries (now only OpenAI)
  "openai>=1.0.0", # OpenAI API client
  "google-generativeai>=0.5.0", 
  "google-api-core>=2.11.0",
  "google-cloud-texttospeech>=2.29.0",
  "langdetect==1.0.9",

  # Utilities and Other Dependencies
  "pyhumps>=2.0.0", # Biblioteca para conversión de camelCase/snake_case
  "pydantic-settings>=2.0.3", # for settings management
  "python-multipart>=0.0.6", # for form data handling
  "fastapi-mail[smtplib]>=1.5.0,<2.0.0", 
  "fastapi-pagination>=0.12.9", # for API pagination
  "fastapi-limiter>=0.1.5", # for rate limiting
  "fastapi-admin>=1.0.4",
  "freezegun>=1.5.4",
  "aiofiles>=23.2.1", # for async file operations
  "boto3>=1.33.9", # AWS SDK if you interact with S3 or other AWS services
  "sentry-sdk[asyncpg,fastapi]>=1.39.1", # error tracking
  "click==8.2.1", # (common CLI tool dependency, often pulled by others)
  "distro==1.9.0", # (system info)
  "PyYAML==6.0.2", # (often used for config)
  "certifi==2025.4.26", # (SSL certificates)
  "charset-normalizer==3.4.2", # (encoding detection)
  "idna==3.10", # (internationalized domain names)
  "urllib3==2.4.0", # (HTTP client dependency)
  "Jinja2==3.1.6", # (templating, often pulled by FastAPI/SQLAdmin)
  "MarkupSafe==3.0.2", # (dependency of Jinja2)
  "uv==0.7.8", # (uv itself can be a dependency for some tools)
  "typing_extensions==4.13.2", # (typing backports)
  "typing-inspection==0.4.1", # (typing inspection)
  "annotated-types==0.7.0", # (pydantic dependency)
  "greenlet==3.2.2", # (greenlet for async/sync bridging, often pulled by SQLAlchemy)
  "anyio==4.9.0", # (async network/file operations, pulled by FastAPI)
  "h11==0.16.0", # (HTTP/1.1 protocol, pulled by Uvicorn/FastAPI)
  "httpcore==1.0.9", # (HTTP core client, pulled by httpx)
  "Pillow==10.1.0", # RE-AÑADIDO: Necesario para 'from PIL import Image'
  "pytz>=2024.1",
  "weasyprint==66.0",
  "resend==2.13.1",

  "py-spy>=0.4.0", # Para perfilado de rendimiento

  # --- Testing ---
  "pytest==8.4.1",
  "pytest-asyncio==0.25.3",
  "pytest-mock==3.12.0",
  "pytest-env==1.1.3",
  "pytest-xdist==3.5.0",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::RuntimeWarning", # Esto es más general y captura las advertencias de SQLAlchemy y Pytest
    "ignore::pytest.PytestUnraisableExceptionWarning:pytest.unraisableexception"
]
asyncio_mode = "auto"
markers = ["integration: Custom mark for integration tests"]
env = ["ENVIRONMENT = test"]

[tool.ruff]
line-length = 120
exclude = ["alembic"]

[tool.ruff.lint]
extend-select = ["W", "E", "I"]
preview = true